revert: changes to flutter engine

+++ engine/src/flutter/impeller/toolkit/interop/paragraph_style.h
@@ -44,7 +44,7 @@ class ParagraphStyle final
 
   void SetTextDecoration(const ImpellerTextDecoration& decoration);
 
-  void SetMaxLines(double max_lines);
+  void SetMaxLines(size_t max_lines);
 
   void SetLocale(std::string locale);
