+++ engine/src/flutter/lib/ui/text.dart
@@ -1905,7 +1905,7 @@ class TextStyle {
 Int32List _encodeParagraphStyle(
   TextAlign? textAlign,
   TextDirection? textDirection,
-  double? maxLines,
+  int? maxLines,
   String? fontFamily,
   double? fontSize,
   double? height,
@@ -2036,7 +2036,7 @@ class ParagraphStyle {
   ParagraphStyle({
     TextAlign? textAlign,
     TextDirection? textDirection,
-    double? maxLines,
+    int? maxLines,
     String? fontFamily,
     double? fontSize,
     double? height,
