+++ engine/src/flutter/lib/web_ui/test/ui/paragraph_style_test.dart
@@ -181,7 +181,7 @@ class _TestParagraphStyleBuilder {
   ui.TextDirection? textDirection;
   ui.FontWeight? fontWeight;
   ui.FontStyle? fontStyle;
-  double? maxLines;
+  int? maxLines;
   String? fontFamily;
   double? fontSize;
   double? height;
